release_date_crawler/requirements.txt
ansicon==1.89.0
attrs==22.2.0
beautifulsoup4==4.12.0
blessed==1.20.0
boto3==1.26.109
botocore==1.29.109
cachetools==5.3.0
certifi==2022.12.7
chalice==1.28.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
inquirer==2.10.1
jinxed==1.2.0
jmespath==1.0.1
packaging==23.0
pluggy==1.0.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytest==7.2.2
pytest-mock==3.10.0
python-dateutil==2.8.2
python-editor==1.0.4
PyYAML==6.0
readchar==4.0.5
requests==2.28.2
rsa==4.9
s3transfer==0.6.0
six==1.16.0
soupsieve==2.4
tomli==2.0.1
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
